refactor(humanoid.js) move humanoid.js to the humanoids directory
