+++ public/js/game/humanoids/humanoid.js
